[REFACTOR] Renamed PatientHealthInfo component as per naming standards
