+++ src/App.js
@@ -4,7 +4,7 @@ import { HashRouter, Route } from "react-router-dom";
 import LoginPage from "../src/pages/LoginPage/LoginPageContainer";
 import ResetPassword from "../src/pages/ResetPassword/ResetPasswordContainer";
 import LandingPage from "../src/pages/LandingPage/LandingPageContainer";
-import PatientHealthInformation from "../src/pages/PatientHealthInfo/PatientHealthInformationContainer";
+import PatientHealthInformation from "./pages/PatientHealthInformation";
 import PrivateRoute from "../src/components/PrivateRoute/PrivateRoute";
 import { makeStyles } from "@material-ui/core/styles";
 import history from "./history";
