+++ src/pages/PatientHealthInformation/index.js
@@ -1,0 +1,3 @@
+import PatientHealthInformation from "./PatientHealthInformation.container";
+
+export default PatientHealthInformation;
